Post Reaction and tags implementation
Implemented the functions for react to a post, in base a the next points:
The user can choose only a like or dislike reaction.
The user can change the reaction of a post.

Besides, the post can be linked with tags during the creation
The get methods for the post return the reactions and tags information.